requirements.txt
certifi==2018.4.16
chardet==3.0.4
idna==2.6
Pillow==5.1.0
reportlab==3.4.0
requests==2.18.4
urllib3==1.22
